refactors and typos
