+++ tobac/tests/test_pbc_utils.py
@@ -2,20 +2,6 @@ import pytest
 import numpy as np
 import tobac.utils.periodic_boundaries as pbc_utils
 import tobac.testing as tb_test
-from collections import Counter
-
-
-def lists_equal_without_order(a, b):
-    """
-    This will make sure the inner list contain the same,
-    but doesn't account for duplicate groups.
-    from: https://stackoverflow.com/questions/31501909/assert-list-of-list-equality-without-order-in-python/31502000
-    """
-    for l1 in a:
-        check_counter = Counter(l1)
-        if not any(Counter(l2) == check_counter for l2 in b):
-            return False
-    return True
 
 
 def test_calc_distance_coords_pbc():
@@ -279,22 +265,22 @@ def test_get_pbc_coordinates():
         (0, 10, 0, 10),
     ]
     # min in hdim_1 and hdim_2
-    assert lists_equal_without_order(
+    assert tb_test.lists_equal_without_order(
         pbc_utils.get_pbc_coordinates(0, 10, 0, 10, -3, 3, -4, 2, "both"),
         [(0, 3, 0, 2), (0, 3, 6, 10), (7, 10, 6, 10), (7, 10, 0, 2)],
     )
     # max in hdim_1, min in hdim_2
-    assert lists_equal_without_order(
+    assert tb_test.lists_equal_without_order(
         pbc_utils.get_pbc_coordinates(0, 10, 0, 10, 5, 12, -4, 2, "both"),
         [(5, 10, 0, 2), (5, 10, 6, 10), (0, 2, 6, 10), (0, 2, 0, 2)],
     )
     # max in hdim_1 and hdim_2
-    assert lists_equal_without_order(
+    assert tb_test.lists_equal_without_order(
         pbc_utils.get_pbc_coordinates(0, 10, 0, 10, 5, 12, 7, 15, "both"),
         [(5, 10, 7, 10), (5, 10, 0, 5), (0, 2, 0, 5), (0, 2, 7, 10)],
     )
     # min in hdim_1, max in hdim_2
-    assert lists_equal_without_order(
+    assert tb_test.lists_equal_without_order(
         pbc_utils.get_pbc_coordinates(0, 10, 0, 10, -3, 3, 7, 15, "both"),
         [(0, 3, 7, 10), (0, 3, 0, 5), (7, 10, 0, 5), (7, 10, 7, 10)],
     )
